Moved date_accessor things to concern
